Refactored Tags module
--HG--
branch : dev

+++ src/Orchard.Web/Modules/Orchard.Tags/Controllers/AdminController.cs
@@ -9,7 +9,6 @@ using Orchard.Settings;
 using Orchard.Tags.Models;
 using Orchard.Tags.ViewModels;
 using Orchard.Tags.Services;
-using Orchard.UI.Notify;
 
 namespace Orchard.Tags.Controllers {
     [ValidateInput(false)]
@@ -22,52 +21,41 @@ namespace Orchard.Tags.Controllers {
             T = NullLocalizer.Instance;
         }
 
-
         public IOrchardServices Services { get; set; }
         protected virtual ISite CurrentSite { get; [UsedImplicitly] private set; }
         
         public Localizer T { get; set; }
 
         public ActionResult Index() {
-            try {
-                IEnumerable<Tag> tags = _tagService.GetTags();
-                var entries = tags.Select(tag => CreateTagEntry(tag)).ToList();
-                var model = new TagsAdminIndexViewModel { Tags = entries };
-                return View(model);
-            }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Listing tags failed: " + exception.Message));
-                return Index();
-            }
+            IEnumerable<Tag> tags = _tagService.GetTags();
+            var entries = tags.Select(CreateTagEntry).ToList();
+            var model = new TagsAdminIndexViewModel { Tags = entries };
+            return View(model);
         }
-
 
         [HttpPost]
         public ActionResult Index(FormCollection input) {
-            var viewModel = new TagsAdminIndexViewModel { Tags = new List<TagEntry>(), BulkAction = new TagAdminIndexBulkAction() };
-            UpdateModel(viewModel);
-
-            try {
-                IEnumerable<TagEntry> checkedEntries = viewModel.Tags.Where(t => t.IsChecked);
-                switch (viewModel.BulkAction) {
-                    case TagAdminIndexBulkAction.None:
-                        break;
-                    case TagAdminIndexBulkAction.Delete:
-                        if (!Services.Authorizer.Authorize(Permissions.ManageTags, T("Couldn't delete tag")))
-                            return new HttpUnauthorizedResult();
-
-                        foreach (TagEntry entry in checkedEntries) {
-                            _tagService.DeleteTag(entry.Tag.Id);
-                        }
-                        break;
-
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
+            var viewModel = new TagsAdminIndexViewModel {Tags = new List<TagEntry>(), BulkAction = new TagAdminIndexBulkAction()};
+            
+            if ( !TryUpdateModel(viewModel) ) {
+                return View(viewModel);
             }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Editing tags failed: " + exception.Message));
-                return Index();
+
+            IEnumerable<TagEntry> checkedEntries = viewModel.Tags.Where(t => t.IsChecked);
+            switch (viewModel.BulkAction) {
+                case TagAdminIndexBulkAction.None:
+                    break;
+                case TagAdminIndexBulkAction.Delete:
+                    if (!Services.Authorizer.Authorize(Permissions.ManageTags, T("Couldn't delete tag")))
+                        return new HttpUnauthorizedResult();
+
+                    foreach (TagEntry entry in checkedEntries) {
+                        _tagService.DeleteTag(entry.Tag.Id);
+                    }
+                    break;
+
+                default:
+                    throw new ArgumentOutOfRangeException();
             }
 
             return RedirectToAction("Index");
@@ -80,67 +68,62 @@ namespace Orchard.Tags.Controllers {
         [HttpPost]
         public ActionResult Create(FormCollection input) {
             var viewModel = new TagsAdminCreateViewModel();
-            try {
-                UpdateModel(viewModel);
-                if (!Services.Authorizer.Authorize(Permissions.CreateTag, T("Couldn't create tag")))
-                    return new HttpUnauthorizedResult();
-                _tagService.CreateTag(viewModel.TagName);
-                return RedirectToAction("Index");
-            }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Creating tag failed: " + exception.Message));
+
+            if (!TryUpdateModel(viewModel)) {
                 return View(viewModel);
             }
+
+            if (!Services.Authorizer.Authorize(Permissions.CreateTag, T("Couldn't create tag")))
+                return new HttpUnauthorizedResult();
+            
+            _tagService.CreateTag(viewModel.TagName);
+            
+            return RedirectToAction("Index");
         }
 
         public ActionResult Edit(int id) {
-            try {
-                Tag tag = _tagService.GetTag(id);
-                var viewModel = new TagsAdminEditViewModel {
-                    Id = tag.Id,
-                    TagName = tag.TagName,
-                };
-                return View(viewModel);
+            Tag tag = _tagService.GetTag(id);
 
+            if(tag == null) {
+                return RedirectToAction("Index");
             }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Retrieving tag information failed: " + exception.Message));
-                return Index();
-            }
+
+            var viewModel = new TagsAdminEditViewModel {
+                Id = tag.Id,
+                TagName = tag.TagName,
+            };
+            
+            return View(viewModel);
         }
 
         [HttpPost]
         public ActionResult Edit(FormCollection input) {
             var viewModel = new TagsAdminEditViewModel();
-            try {
-                UpdateModel(viewModel);
-                if (!Services.Authorizer.Authorize(Permissions.ManageTags, T("Couldn't edit tag")))
-                    return new HttpUnauthorizedResult();
 
-                _tagService.UpdateTag(viewModel.Id, viewModel.TagName);
-                return RedirectToAction("Index");
-            }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Editing tag failed: " + exception.Message));
+            if ( !TryUpdateModel(viewModel) ) {
                 return View(viewModel);
             }
+            
+            if (!Services.Authorizer.Authorize(Permissions.ManageTags, T("Couldn't edit tag")))
+                return new HttpUnauthorizedResult();
+
+            _tagService.UpdateTag(viewModel.Id, viewModel.TagName);
+            return RedirectToAction("Index");
         }
 
         public ActionResult Search(int id) {
-            try {
-                Tag tag = _tagService.GetTag(id);
-                IEnumerable<IContent> contents = _tagService.GetTaggedContentItems(id).ToList();
-                var viewModel = new TagsAdminSearchViewModel {
-                    TagName = tag.TagName,
-                    Contents = contents,
-                };
-                return View(viewModel);
+            Tag tag = _tagService.GetTag(id);
 
+            if (tag == null) {
+                return RedirectToAction("Index");
             }
-            catch (Exception exception) {
-                Services.Notifier.Error(T("Retrieving tagged items failed: " + exception.Message));
-                return Index();
-            }
+
+            IEnumerable<IContent> contents = _tagService.GetTaggedContentItems(id).ToList();
+            var viewModel = new TagsAdminSearchViewModel {
+                TagName = tag.TagName,
+                Contents = contents,
+            };
+            return View(viewModel);
         }
 
         private static TagEntry CreateTagEntry(Tag tag) {
